# back/docs/api.yml
openapi: 3.0.0
paths:
...
  /schedule:
    get:
      operationId: ScheduleController_search
      parameters:
        - name: page
          required: false
          in: query
          description: The page number
          schema:
            type: number
        - name: perPage
          required: false
          in: query
          description: The number of items per page
          schema:
            type: number
        - name: sort
          required: false
          in: query
          description: The field that should be used for sorting
          schema:
            enum:
              - createdAt
            type: string
        - name: sortDir
          required: false
          in: query
          description: The sort direction
          schema:
            enum:
              - asc
              - desc
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: object
                properties:
                  meta:
                    type: object
                    properties:
                      totalItems:
                        type: number
                      itemCount:
                        type: number
                      itemsPerPage:
                        type: number
                      totalPages:
                        type: number
                      currentPage:
                        type: number
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/SchedulePresenter'
        '401':
          description: Unauthorized
        '422':
          description: Unprocessable Entity
      tags:
        - schedule
  /schedule/{id}:
    get:
      operationId: ScheduleController_findOne
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '401':
          description: Unauthorized
        '404':
          description: Schedule not found
      tags:
        - schedule
      security:
        - bearer: []
    put:
      operationId: ScheduleController_update
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateScheduleDto'
      responses:
        '401':
          description: Unauthorized
        '404':
          description: Schedule not found
        '422':
          description: Unprocessable Entity
      tags:
        - schedule
      security:
        - bearer: []
    delete:
      operationId: ScheduleController_remove
      parameters:
        - name: id
          required: true
          in: path
          schema:
            type: string
      responses:
        '204':
          description: Schedule deleted
        '401':
          description: Unauthorized
        '404':
          description: Schedule not found
      tags:
        - schedule
      security:
        - bearer: []
...
info:
  title: Nest Auth Api
  description: The Nest Api support creating deleting editing and authenticating users
  version: 1.0.0
  contact: {}
...
servers: []
components:
  securitySchemes:
    bearer:
      scheme: Bearer
      bearerFormat: JWT
      name: Authorization
      description: JWT token
      in: Header
      type: http
  schemas:
...
    UpdateScheduleDto:
      type: object
      properties:
        classroomId:
          type: string
          description: The new classroomId of the schedule
        disciplineId:
          type: string
          description: The new disciplineId of the schedule
      required:
        - classroomId
        - disciplineId
...